eduqgen/requirements.txt
flask
tensorflow
numpy
pandas
scikit-learn
transformers
torch
sentencepiece
pymupdf
pytesseract
Pillow
nltk
datasets
